Create Bisection Method
Problem 2. Write and test a program to implement the bisection method. Test the program
on the following functions and intervals
(a) x^-1 - tan(x) on [0, 22/14].
(b) x^-1- 2^x on [0, 1].
(c) Find a root of
x^8 - 36x^7 + 546x^6 - 4536x^5 + 22449x^4 - 67284x^3 + 118124x^2 - 109584x + 40320 = 0
in the interval [5.5,6.5].
(d) In (c), change 36 to 36.001 and repeat the test.